Struct roommates::sharing::SharedBill

pub struct SharedBill { /* fields omitted */ }

A Bill along with the communally shared cost

Bills can be transformed directly into SharedBills (with the Bills fixed cost becoming the shared amount) but for bills that are usage dependent, there might be an additional implicit fixed cost that can be inferred based on a bill history

Implementations

impl SharedBill

pub fn new(bill: Bill, shared_amount: Money) -> Result<Self, Error>

Creates a new SharedBill

Examples

use roommates::{DateInterval, sharing::{Bill, SharedBill}};
use steel_cent::{Money, currency::USD};

let water_bill = Bill::new(
    Money::of_minor(USD, 83_22),
    DateInterval::new((2020, 4, 15), (2020, 5, 15)).unwrap(),
);
let water_bill = SharedBill::new(water_bill, Money::of_minor(USD, 30_00));
assert!(water_bill.is_ok());

pub fn shared_amount(&self) -> Money

Returns the amount that should be divided evenly

Examples

use roommates::{DateInterval, sharing::{Bill, SharedBill}};
use steel_cent::{Money, currency::USD};

let water_bill = Bill::new(
    Money::of_minor(USD, 83_22),
    DateInterval::new((2020, 4, 15), (2020, 5, 15)).unwrap(),
);
let water_bill = SharedBill::new(water_bill, Money::of_minor(USD, 30_00)).unwrap();
assert_eq!(water_bill.shared_amount(), Money::of_minor(USD, 30_00));

pub fn amount_due(&self) -> Money

Returns the total cost

Examples

use roommates::{DateInterval, sharing::{Bill, SharedBill}};
use steel_cent::{Money, currency::USD};

let water_bill = Bill::new(
    Money::of_minor(USD, 83_22),
    DateInterval::new((2020, 4, 15), (2020, 5, 15)).unwrap(),
);
let water_bill = SharedBill::new(water_bill, Money::of_minor(USD, 30_00)).unwrap();
assert_eq!(water_bill.amount_due(), Money::of_minor(USD, 83_22));

pub fn usage_period(&self) -> DateInterval

Returns the period of time for which is being billed

Examples

use roommates::{DateInterval, sharing::{Bill, SharedBill}};
use steel_cent::{Money, currency::USD};

let water_bill = Bill::new(
    Money::of_minor(USD, 83_22),
    DateInterval::new((2020, 4, 15), (2020, 5, 15)).unwrap(),
);
let water_bill = SharedBill::new(water_bill, Money::of_minor(USD, 30_00)).unwrap();
assert_eq!(
    water_bill.usage_period(),
    DateInterval::new((2020, 4, 15), (2020, 5, 15)).unwrap(),
);

impl SharedBill

pub fn from_estimate<'a, I>(
bill_with_notes: (Bill, (u32, Option<f64>)),
bill_history: I
) -> Result<Self, Box<dyn Error>> where
I: IntoIterator<Item = (&'a Bill, u32, Option<f64>)>,

builds a linear model to predict amount due based on bill history, then applies the model to the current month, setting the occupancy to zero

Performs poorly if temperature index and occupancy always vary together

You will need several bills worth of data for this to work

pub fn from_fixed(bill: Bill) -> Self